services/admin_service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.30.0
httpx>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
pyjwt>=2.8.0
bcrypt>=4.1.0
structlog>=24.1.0
prometheus-client>=0.20.0
